
\name{size-internal}

\title{Internal size Objects}
% terms

\alias{HPD.density}
\alias{HTestimator}
\alias{Zee}
\alias{akestN}
\alias{allcmp}
\alias{beginparallel}
\alias{bins}
\alias{bnwnest}
\alias{cmp.mu}
\alias{cmp.natural}
\alias{cmpmle}
\alias{dcmp}
\alias{dcmp.natural}
\alias{endparallel}
\alias{getest}
\alias{getestC}
\alias{getestCstacked}
\alias{getincl}
\alias{getincl.raw}
\alias{getinclC}
\alias{getinclCstacked}
\alias{ldcmp.natural}
\alias{likcmp}
\alias{likelihoodsize}
\alias{ll.density}
\alias{lllspps}
\alias{llspps}
\alias{mleNest}
\alias{mode.density}
\alias{poscmp}
\alias{posize.warning}
\alias{posnbinom}
\alias{pospln}
\alias{ppspolya}
\alias{probtodist}
\alias{robust.inverse}
\alias{roundstoc}
\alias{sppsestN}
\alias{sppssample}
\alias{sppssamplei}

%
% These need to be documented!
%
\alias{impute.size}
\alias{posteriornbinom}
\alias{priorsize}

\description{
  Internal size functions.
}
% \usage{
%anova.size(object, \dots)
%}

\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
%
%\arguments{
% 
%\item{\dots}{further arguments passed to or used by methods.}
%}

\seealso{size, statnet-package}

\keyword{internal}
