% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorsize.R
\name{posize_warning}
\alias{posize_warning}
\title{Warning message for posteriorsize fit failure}
\usage{
posize_warning()
}
\value{
\code{\link{posize_warning}} returns a character string witn the warning message.
}
\description{
\code{\link{posteriorsize}} computes the posterior distribution of the
population size based on data collected by Respondent Driven Sampling.
This function returns the warning message if it fails. 
It enables packages that call \code{\link{posteriorsize}} to use
a consistent error message.
}
\seealso{
posteriorsize
}
\keyword{models}
