% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.sspse.R
\name{print.summary.sspse}
\alias{print.summary.sspse}
\title{Summarizing Population Size Estimation Model Fits}
\usage{
\method{print}{summary.sspse}(
  x,
  digits = max(3, getOption("digits") - 3),
  correlation = FALSE,
  covariance = FALSE,
  signif.stars = getOption("show.signif.stars"),
  eps.Pvalue = 1e-04,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"summary.sspse"}, usually, a result of a
call to \code{summary.sspse}.}

\item{digits}{the number of significant digits to use when printing.}

\item{correlation}{logical; if \code{TRUE}, the correlation matrix of the
estimated parameters is returned and printed.}

\item{covariance}{logical; if \code{TRUE}, the covariance matrix of the
estimated parameters is returned and printed.}

\item{signif.stars}{logical. If \code{TRUE}, \sQuote{significance stars} are
printed for each coefficient.}

\item{eps.Pvalue}{number; indicates the smallest p-value. 
\code{\link[stats]{printCoefmat}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.sspse} computes and returns a two row matrix of
summary statistics of the prior and estimated posterior distributions. 
The rows correspond to the \code{Prior} and the \code{Posterior}, respectively. 
The rows names are \code{Mean}, \code{Median}, \code{Mode}, 
\code{25\%}, \code{75\%}, and \code{90\%}.
These correspond to the distributional mean, median, mode, lower quartile,
upper quartile and 90\% quantile, respectively.
}
\description{
This is the \code{print} method for the \code{summary} class method for class \code{"sspse"} objects.
These objects encapsulate an estimate of the posterior distribution of
the population size based on data collected by Respondent Driven Sampling.
The approach approximates the RDS via the Sequential Sampling model of Gile
(2008).  As such, it is referred to as the Sequential Sampling - Population Size Estimate (SS-PSE).
It uses the order of selection of the sample to provide information
on the distribution of network sizes over the population members.
}
\details{
\code{print.summary.sspse} tries to be smart about formatting the
coefficients, standard errors, etc. and additionally gives
\sQuote{significance stars} if \code{signif.stars} is \code{TRUE}.

Aliased coefficients are omitted in the returned object but restored by the
\code{print} method.

Correlations are printed to two decimal places (or symbolically): to see the
actual correlations print \code{summary(object)$correlation} directly.
}
\examples{

data(fauxmadrona)
# Here interval=1 so that it will run faster. It should be higher in a 
# real application.
fit <- posteriorsize(fauxmadrona, median.prior.size=1000,
                                 burnin=20, interval=1, samplesize=100)
fit

}
\references{
Gile, Krista J. (2008) \emph{Inference from Partially-Observed Network
Data}, Ph.D. Thesis, Department of Statistics, University of Washington.

Gile, Krista J. and Handcock, Mark S. (2010) \emph{Respondent-Driven
Sampling: An Assessment of Current Methodology}, Sociological Methodology
40, 285-327.

Gile, Krista J. and Handcock, Mark S. (2014) \pkg{sspse}: Estimating Hidden 
Population Size using Respondent Driven Sampling Data
R package, Los Angeles, CA.  Version 0.5, \url{https://hpmrg.org/sspse/}.

Handcock MS (2003).  \pkg{degreenet}: Models for Skewed Count Distributions
Relevant to Networks.  Statnet Project, Seattle, WA.  Version 1.2,
\url{https://statnet.org/}.

Handcock, Mark S., Gile, Krista J. and Mar, Corinne M. (2014)
\emph{Estimating Hidden Population Size using Respondent-Driven Sampling
Data}, Electronic Journal of Statistics, 8, 1, 1491-1521

Handcock, Mark S., Gile, Krista J. and Mar, Corinne M. (2015)
\emph{Estimating the Size of Populations at High Risk for HIV using Respondent-Driven 
Sampling Data}, Biometrics.
}
\seealso{
The model fitting function \code{\link{posteriorsize}},
\code{\link{summary}}.

Function \code{\link[stats]{coef}} will extract the matrix of coefficients with
standard errors, t-statistics and p-values.
}
\keyword{models}
